NumFracDiff

Documentation for NumFracDiff.

NumFracDiff.compute!

NumFracDiff.generate_weights!

NumFracDiff.init_workspace

NumFracDiff.optimal_L!

NumFracDiff.compute! — Method

compute!(method, ws::NumDiffWorkspace, data::Vector{NumDiffFloat}, prob::NumDiffProblem)

Compute the Riemann-Liouville fractional derivative of the input signal `data` and store the result
directly in `ws.deriv`. This function has two variants depending on `method`:

1. **Serial version** (`method::RL`) 
2. **Parallel version** (`method::RLThreads`) 

# Arguments

- `method::RL` or `RLThreads`: Method descriptor. Must contain field `L::NumDiffInt` specifying 
the short-memory length.  
- `ws::NumDiffWorkspace`: Workspace containing precomputed `weights` and the derivative vector `deriv`.  
The derivative will be written **in-place** to avoid additional memory allocations.  
- `data::Vector{NumDiffFloat}`: Input signal or time series to differentiate.  
- `prob::NumDiffProblem`: Problem definition containing the derivative order `order`, time step `dt`, 
and signal length `n`.

# Returns

- `ws.deriv` updated **in-place** with the computed fractional derivative.

NumFracDiff.generate_weights! — Method

generate_weights!(::RL, prob::NumDiffProblem, ws::NumDiffWorkspace)

Compute the convolution weights for the Riemann–Liouville (RL) fractional derivative
and store them **in-place** in `ws.weights`.

# Arguments
- `::RL` : dispatch type for RL methods Union{RL, RLThreads, RLShortMemCorr, RLShortMemCorrThreads}
- `prob::NumDiffProblem` : numerical differentiation problem definition
- `ws::NumDiffWorkspace` : workspace containing preallocated `weights` vector (length ≥ `prob.n`)

NumFracDiff.init_workspace — Method

init_workspace(prob::NumDiffProblem) -> NumDiffWorkspace

Create a reusable workspace for numerical fractional differentiation.

Precomputes convolution weights based on the given problem `prob` and
allocates a buffer to store the derivative of a signal of length `prob.n`.

The returned workspace should be reused across calls to `compute!`
to **avoid repeated allocations** and improve performance, especially
in loops or optimization routines.

# Arguments
- `prob::NumDiffProblem` : Numerical differentiation problem definition.

# Returns
A `NumDiffWorkspace` containing:
- `weights` : Precomputed convolution weights.
- `deriv` : Preallocated vector for storing the derivative.

NumFracDiff.optimal_L! — Method

optimal_L(data, prob; tol=1e-2)

Compute optimal memory length L (in samples) for the Short-Memory method.

# Arguments

- data : input signal
- prob : NumDiffProblem (must contain dt and order)
- tol  : tolerated tail error
- safety : safety factor (>1 increases robustness)

Returns
-------
Integer memory length (number of time steps).